Add progress bar in verbose mode, fix performance issue
As per Peter, generating constraints on the accumulating bottom-up graph
is unnecessary, since sketch generation inlines callee's sketches. This
seems to be true based on my testing and a quick look over the code.

This is like a 20x speedup on a medium-sized binary (like monkey http).

I also added a verbose-mode to the Solver which uses tqdm, and cleaned
up dead Python requirements.