Fix regular language violations on output constraints
As per the paper, we only want to accept paths of the language (recall
_)* (forget _)*

Also added a field to track the sources of a type, in sketch nodes. This
will allow clients to do fancier things if they want to do some type
unification when emitting C types.

Also fixed a bug where the global graph was not using
DerivedTypeVariables (just strings) for the keys.