Fix global handling
On the negative side, this has a 2.5x slow-down and an almost 2x RAM
overhead from the previous version.
On the plus side, the results are a lot more correct. After
experimenting with a few different ways of doing globals and not losing
information, this is the most performant and the simplest.