Fix typevar handling in sketch gen
We need to use the node from the interprocedural sketch to meet or join
the atomic type, otherwise we lose information. Unfortunately we need to
traverse a lot more nodes this way, which makes the performance a lot
worse.
Better to have worse performance (approx. 2x slower on real stuff) than
missing types, in this case. Hopefully we can find a way to speed it
back up again later.